src/Server/requirements/local.txt
django-debug-toolbar==3.2
django-extensions==3.1.0
Werkzeug==1.0.1